Core/Areatrigger Fix spawn_mask unsupported difficulty error

+++ src/server/game/Globals/AreaTriggerDataStore.h
@@ -30,7 +30,7 @@ struct AreaTriggerData
     float position_y;
     float position_z;
     uint32 map_id;
-    uint32 spawn_mask;
+    uint64 spawn_mask;
     uint32 scriptId;
     ObjectGuid::LowType guid;
 };
